PSR-0
 - Restructured the lib dir into src
 - Moved httpful.php into PSR-0 file hierarchy
 - Added a basic autoloader via spl
 - Added bootstrap file
