+++ bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+
+require(__DIR__ . '/src/Httpful/Bootstrap.php');
+\Httpful\Bootstrap::init();
