+++ examples/showclix.php
@@ -1,6 +1,6 @@
 <?php
 
-require(__DIR__ . '/../lib/httpful.php');
+require(__DIR__ . '/../bootstrap.php');
 
 // Get event details for a public event
 $uri = "http://api.showclix.com/Event/8175";
@@ -9,4 +9,4 @@ $response = \Httpful\Request::get($uri)
     ->sendIt();
 
 // Print out the event details
-echo "The event {$response->body->event} will take place on {$response->body->event_start}";
+echo "The event {$response->body->event} will take place on {$response->body->event_start}\n";
